feat: update logic, update mdoal styles, add 'bcc' option in sending message
