+++ src/components/Modal/Modal.module.scss
@@ -46,13 +46,39 @@
   span {
     font-weight: 700;
   }
+
+  animation: modal_load 0.3s;
 }
 
-.modal_module_header{
+.modal_module_field_button {
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+
+  padding: 0 1rem;
+  margin: 0;
+
+  display: flex;
+  flex-direction: row-reverse;
+  div {
+    width: fit-content;
+    cursor: pointer;
+    color: #99a1b3;
+  }
+  div:hover {
+    color: #b2b4bd;
+  }
+  div:active {
+    color: #d1d1ec;
+  }
+}
+
+.modal_module_header {
   padding: 15px;
   border-bottom: 1px solid #e5e5e5;
 }
-.modal_module_footer{
+.modal_module_footer {
   border-top: 1px solid #e5e5e5;
   padding: 15px;
   display: flex;
