+++ src/models/SendMessageData.ts
@@ -1,6 +1,7 @@
 export interface sendMessageData {
   to?: string;
   cc?: string;
+  bcc?: string;
   subject?: string;
   messageText?: string;
 }
