+++ src/styles/App.module.scss
@@ -1,0 +1,20 @@
+.modal_open_button {
+  width: 6rem;
+  height: 3rem;
+  border: none;
+  background: #fff;
+  color: #646d81;
+  font-size: 14px;
+  margin: 0 2rem;
+  font-weight: 500;
+  border-radius: 0.2rem;
+}
+
+.modal_open_button:hover {
+  box-shadow: inset 0 0.2rem 0.2rem 0 rgba(0, 0, 0,0.5);
+}
+
+.modal_open_button:active {
+  box-shadow: inset 0 0.2rem 0.2rem 0 rgba(0, 0, 0,1);
+
+}
